This is my initial upload
In this initial upload I am adding the files that create the base for my project. I believe I am in a spot were I have completed the requirements for milestone 1. There are many elements of this project that I am planning on changing and adding. The login system is functioning only with 1 user at the moment - it will be expanded to accommodate multiple users.